+++ DeepFlight/src/rendering/DrawableTexture.cs
@@ -14,11 +14,11 @@ public class DrawableTexture : Movable {
         Texture = texture;
     }
 
-    public DrawableTexture(Texture2D texture, int width, int height) : base(width, height ) {
+    public DrawableTexture(Texture2D texture, float width, float height) : base(width, height ) {
         Texture = texture;
     }
 
-    public DrawableTexture(Texture2D texture, Color col, int width, int height, double x, double y) : base(width, height) {
+    public DrawableTexture(Texture2D texture, Color col, float width, float height, double x, double y) : base(width, height) {
         Texture = texture;
         X = x;
         Y = y;
